refactor(App): adjust to manage useSignOut and logout when generate a error
