+++ src/pages/error/type.ts
@@ -1,0 +1,6 @@
+interface IOptionError {
+  code: number;
+  signOut: boolean;
+}
+
+export type { IOptionError };
